requirement.txt
amqp==1.4.9
ansible==2.8.5
anyjson==0.3.3
APScheduler==3.6.1
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-celery==3.3.1
django-simpleui==3.2
djangorestframework==3.10.3
Jinja2==2.10.1
kombu==3.0.37
MarkupSafe==1.1.1
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
tzlocal==2.0.0
uWSGI==2.0.18
